+++ src/test/java/science/icebreaker/device_availability/DeviceAvailabilityServiceTest.java
@@ -17,8 +17,8 @@ import science.icebreaker.account.AccountRepository;
 import science.icebreaker.account.AccountService;
 import science.icebreaker.account.RegistrationRequest;
 import science.icebreaker.account.RegistrationRequestMock;
-import science.icebreaker.device_availability.Exceptions.DeviceAvailabilityCreationException;
-import science.icebreaker.device_availability.Exceptions.DeviceAvailabilityNotFoundException;
+import science.icebreaker.exception.DeviceAvailabilityNotFoundException;
+import science.icebreaker.exception.DeviceAvailabilityCreationException;
 import science.icebreaker.mail.MailException;
 import science.icebreaker.wiki.WikiPage;
 import science.icebreaker.wiki.WikiPageRepository;
